--- src/main/java/QueryBuilderMoc/QueryBuilder.java
@@ -1,7 +1,0 @@
-package QueryBuilderMoc;
-
-//Not needed for now?
-public interface QueryBuilder {
-
-  String getQueryContent();
-}
