--- src/main/java/QueryBuilderMoc/QueryBuilderCQL.java
@@ -1,9 +1,0 @@
-package QueryBuilderMoc;
-
-public class QueryBuilderCQL implements QueryBuilder {
-
-  @Override
-  public String getQueryContent() {
-    return "CQLQuery";
-  }
-}
