--- src/main/java/QueryBuilderMoc/QueryBuilderFHIR.java
@@ -1,9 +1,0 @@
-package QueryBuilderMoc;
-
-public class QueryBuilderFHIR implements QueryBuilder {
-
-  @Override
-  public String getQueryContent() {
-    return "FHIRQuery";
-  }
-}
