--- src/main/java/QueryExecutor/api/QueryNotFoundException.java
@@ -1,10 +1,0 @@
-package QueryExecutor.api;
-
-/**
- * Indicates that a requested query was not found.
- */
-public class QueryNotFoundException extends Exception {
-    public QueryNotFoundException(String queryId) {
-        super("Query with ID '" + queryId + "' could not be found.");
-    }
-}
