--- src/main/java/QueryExecutor/api/QueryStatus.java
@@ -1,12 +1,0 @@
-package QueryExecutor.api;
-
-/**
- * Represents different states a query can be in.
- */
-public enum QueryStatus {
-  RETRIEVED,
-  QUEUED,
-  EXECUTING,
-  COMPLETED,
-  FAILED
-}
