+++ src/main/java/de/numcodex/feasibility_gui_backend/model/query/StructuredQuery.java
@@ -1,0 +1,23 @@
+package de.numcodex.feasibility_gui_backend.model.query;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import de.numcodex.feasibility_gui_backend.model.common.Criterion;
+import lombok.Data;
+
+import java.net.URI;
+import java.util.List;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class StructuredQuery {
+
+  @JsonProperty
+  private URI version;
+  @JsonProperty("inclusionCriteria")
+  private List<List<Criterion>> inclusionCriteria;
+  @JsonProperty("exclusionCriteria")
+  private List<List<Criterion>> exclusionCriteria;
+  @JsonProperty("display")
+  private String display;
+}
