+++ src/main/java/de/numcodex/feasibility_gui_backend/service/query_builder/QueryBuilder.java
@@ -1,0 +1,9 @@
+package de.numcodex.feasibility_gui_backend.service.query_builder;
+
+import de.numcodex.feasibility_gui_backend.model.query.StructuredQuery;
+
+//Not needed for now?
+public interface QueryBuilder {
+
+  String getQueryContent(StructuredQuery query);
+}
