+++ src/main/java/de/numcodex/feasibility_gui_backend/service/query_builder/QueryBuilderCQL.java
@@ -1,0 +1,11 @@
+package de.numcodex.feasibility_gui_backend.service.query_builder;
+
+import de.numcodex.feasibility_gui_backend.model.query.StructuredQuery;
+
+public class QueryBuilderCQL implements QueryBuilder {
+
+  @Override
+  public String getQueryContent(StructuredQuery query) {
+    return "CQLQuery";
+  }
+}
