+++ src/main/java/de/numcodex/feasibility_gui_backend/service/query_executor/impl/dsf/DSFQueryResult.java
@@ -1,0 +1,17 @@
+package de.numcodex.feasibility_gui_backend.service.query_executor.impl.dsf;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * Holds query result information.
+ *
+ * The information is related to a specific query (queryId) and an organization that the result is coming from (siteId).
+ */
+@Data
+@AllArgsConstructor
+class DSFQueryResult {
+    private final String queryId;
+    private final String siteId;
+    private final int measureCount;
+}
