+++ src/test/java/de/numcodex/feasibility_gui_backend/FeasabilityGuiBackendApplicationTests.java
@@ -1,16 +1,17 @@
 package de.numcodex.feasibility_gui_backend;
 
-import QueryBuilderMoc.QueryBuilderCQL;
-import QueryBuilderMoc.QueryBuilderFHIR;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.numcodex.feasibility_gui_backend.model.query.QueryDefinition;
+import de.numcodex.feasibility_gui_backend.model.query.StructuredQuery;
 import de.numcodex.feasibility_gui_backend.model.ui.TerminologyEntry;
-import de.numcodex.feasibility_gui_backend.service.QueryBuilderService;
-import java.net.URI;
-import java.net.URL;
+import de.numcodex.feasibility_gui_backend.service.QueryHandlerService;
+import de.numcodex.feasibility_gui_backend.service.query_builder.QueryBuilderCQL;
+import de.numcodex.feasibility_gui_backend.service.query_builder.QueryBuilderFHIR;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import java.net.URI;
+import java.net.URL;
 
 
 class FeasibilityGuiBackendApplicationTests {
@@ -21,16 +22,16 @@ class FeasibilityGuiBackendApplicationTests {
 
     objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
 
-    QueryDefinition queryDefinition = objectMapper.readValue(
+    StructuredQuery structuredQuery = objectMapper.readValue(
         new URL("file:src/test/resources/2021_02_01_QueryExampleWithoutTimeRestriction.json"),
-        QueryDefinition.class);
+        StructuredQuery.class);
     Assertions.assertEquals(new URI("http://to_be_decided.com/draft-1/schema#"),
-        queryDefinition.getVersion());
+        structuredQuery.getVersion());
   }
 
   @Test
   public void queryBuilder() {
-    var queryService = new QueryBuilderService();
+    var queryService = new QueryHandlerService();
     var cqlBuilder = new QueryBuilderCQL();
     var fhirBuilder = new QueryBuilderFHIR();
     Assertions.assertEquals("FHIRQuery", queryService.getQueryContent(fhirBuilder));
